+++ private/rdf4cpp/parser/JsonLdParser.hpp
@@ -6,6 +6,8 @@
 #include <rdf4cpp/Quad.hpp>
 #include <rdf4cpp/parser/IStreamQuadIterator.hpp>
 
+#include <rdf4cpp/parser/JsonLdParserTypes.hpp>
+
 #include <generator>
 #include <vector>
 #include <forward_list>
@@ -13,240 +15,28 @@
 #include <simdjson.h>
 
 namespace rdf4cpp::parser {
-    static constexpr std::string_view keyword_base = "@base";
-    static constexpr std::string_view keyword_container = "@container";
-    static constexpr std::string_view keyword_context = "@context";
-    static constexpr std::string_view keyword_direction = "@direction";
-    static constexpr std::string_view keyword_graph = "@graph";
-    static constexpr std::string_view keyword_id = "@id";
-    static constexpr std::string_view keyword_import = "@import";
-    static constexpr std::string_view keyword_included = "@included";
-    static constexpr std::string_view keyword_index = "@index";
-    static constexpr std::string_view keyword_json = "@json";
-    static constexpr std::string_view keyword_language = "@language";
-    static constexpr std::string_view keyword_list = "@list";
-    static constexpr std::string_view keyword_nest = "@nest";
-    static constexpr std::string_view keyword_none = "@none";
-    static constexpr std::string_view keyword_prefix = "@prefix";
-    static constexpr std::string_view keyword_propagate = "@propagate";
-    static constexpr std::string_view keyword_protected = "@protected";
-    static constexpr std::string_view keyword_reverse = "@reverse";
-    static constexpr std::string_view keyword_set = "@set";
-    static constexpr std::string_view keyword_type = "@type";
-    static constexpr std::string_view keyword_value = "@value";
-    static constexpr std::string_view keyword_version = "@version";
-    static constexpr std::string_view keyword_vocab = "@vocab";
-
-    // not treated as keyword in other places
-    static constexpr std::string_view keyword_default = "@default";
-
-    static constexpr std::string_view rdf_json_datatype = "http://www.w3.org/1999/02/22-rdf-syntax-ns#JSON";
-    static constexpr std::string_view iri_nil = "http://www.w3.org/1999/02/22-rdf-syntax-ns#nil";
-    static constexpr std::string_view iri_rest = "http://www.w3.org/1999/02/22-rdf-syntax-ns#rest";
-    static constexpr std::string_view iri_first = "http://www.w3.org/1999/02/22-rdf-syntax-ns#first";
-
-    namespace json_ld {
-        struct Null {
-            constexpr auto operator<=>(const Null &) const noexcept = default;
-        };
-        struct NotSet {
-            constexpr auto operator<=>(const NotSet &) const = default;
-        };
-
-        enum class BaseDirection : uint8_t {
-            None,
-            Ltr,
-            Rtl,
-        };
-        enum class ParseState : uint8_t {
-            NotStarted,
-            InProgress,
-            Done,
-        };
-
-        enum class IRIMappingType : uint8_t {
-            None,
-            IRI,
-            BlankNode,
-            Keyword,
-        };
-        struct IRIMapping {
-            std::string data;
-            IRIMappingType type = IRIMappingType::None;
-            std::string search_key = "";
-
-            constexpr auto operator<=>(IRIMapping const &r) const noexcept {
-                return std::tie(data, type) <=> std::tie(r.data, r.type);
-            }
-            constexpr bool operator==(IRIMapping const &r) const noexcept {
-                return (*this <=> r) == std::strong_ordering::equal;
-            }
-
-            [[nodiscard]] constexpr bool is_keyword(std::string_view k) const {
-                return type == IRIMappingType::Keyword && data == k;
-            }
-            [[nodiscard]] constexpr std::string_view get_search_key() const {
-                if (search_key.empty()) {
-                    return data;
-                }
-                return search_key;
-            }
-        };
-        struct TypedLiteralMapping {
-            std::string value;
-            std::string type;
-        };
-        struct LiteralMapping {
-            std::string value;
-            std::optional<std::string> language;
-            BaseDirection direction;
-        };
-        using ExpandedValue = std::variant<IRIMapping, LiteralMapping, TypedLiteralMapping, simdjson::ondemand::value>;
-        struct LanguageMapping : std::variant<NotSet, Null, std::string> {
-            using Base = std::variant<NotSet, Null, std::string>;
-
-            using Base::Base;
-            using Base::operator=;
-            constexpr auto operator<=>(const LanguageMapping &) const = default;
-
-            constexpr LanguageMapping const &operator||(LanguageMapping const & other) const noexcept {
-                if (std::holds_alternative<NotSet>(*this)) {
-                    return other;
-                }
-                return *this;
-            }
-
-            [[nodiscard]] constexpr std::optional<std::string> output() const noexcept {
-                return std::visit([]<typename T>(const T& e) -> std::optional<std::string> {
-                    if constexpr (std::same_as<T, std::string>) {
-                        return e;
-                    }
-                    else {
-                        return std::nullopt;
-                    }
-                }, *this);
-            }
-        };
-
-        struct TermDefinitionBase {
-            std::string key;
-            IRIMapping iri_mapping;
-            bool is_prefix = false;
-            bool is_reverse_property = false;
-            bool is_simple = false;
-            std::optional<std::string> base_iri;
-            std::optional<std::string> context;
-            std::vector<std::string> container_mapping;
-            IRIMapping index_mapping;
-            LanguageMapping language_mapping = NotSet{};
-            std::optional<std::string> nest_value;
-            std::optional<std::string> type_mapping;
-            BaseDirection direction_mapping = BaseDirection::None;
-
-            constexpr auto operator<=>(TermDefinitionBase const &) const = default;
-
-            constexpr explicit TermDefinitionBase(std::string_view k)
-                : key{k} {
-            }
-
-            [[nodiscard]] constexpr bool has_container_mapping(std::string_view m) const {
-                for (const auto& e : container_mapping) {
-                    if (m == e) {
-                        return true;
-                    }
-                }
-                return false;
-            }
-        };
-        struct TermDefinition : TermDefinitionBase {
-            bool is_protected = false;
-            bool needs_context_check = false;
-            ParseState parse_state = ParseState::NotStarted;
-
-            constexpr explicit TermDefinition(std::string_view k)
-                : TermDefinitionBase{k} {
-            }
-        };
-
-        struct Context {
-            std::vector<TermDefinition> terms{};  // TODO map? one of the term members as key
-            std::string base_iri;
-            std::optional<std::string> vocab = std::nullopt;
-            LanguageMapping language = NotSet{};
-            BaseDirection base_direction = BaseDirection::None;
-            Context const* previous_context = nullptr;
-
-            TermDefinition *try_find_term(std::string_view key);
-            [[nodiscard]] TermDefinition const *try_find_term(std::string_view key) const;
-        };
-
-        struct KeyPath {
-            std::vector<std::variant<std::string, size_t>> keys;
-        };
-        struct ExpandedMapEntry;
-        struct ExpandedMap {
-            std::vector<ExpandedMapEntry> entries;
-            Context* active_context = nullptr;
-            std::forward_list<Context> context_storage; // moving the contained objects is not allowed
-            bool expanded_from_no_map = false;
-
-            constexpr ExpandedMapEntry* try_find_entry(IRIMapping const &key);
-            constexpr ExpandedMapEntry* try_find_keyword(std::string_view key);
-        };
-        using ExpandedLevel = std::variant<ExpandedMap, LiteralMapping, TypedLiteralMapping, Null, simdjson::ondemand::value, simdjson::ondemand::array>;
-
-
-        struct ExpandedMapEntry {
-            IRIMapping key;
-            KeyPath path;
-            std::vector<IRIMapping> keyword_values;
-            std::string active_property = "";
-            bool is_json_literal = false;
-            bool as_list = false;
-            bool as_graph = false;
-            bool is_reverse = false;
-            bool as_multiple_graphs = false;
-            std::optional<IRIMapping> as_named_graph = std::nullopt;
-            std::optional<BaseDirection> language_map = std::nullopt;
-            std::optional<ExpandedValue> pre_expanded_value = std::nullopt;
-            Context const *active_context = nullptr;
-            std::optional<ExpandedLevel> next_level_pre_expanded = std::nullopt;
-
-            constexpr bool has_keyword_value(std::string_view key) {
-                return std::ranges::any_of(keyword_values, [&](auto const& t) { return t.type == IRIMappingType::Keyword && t.data == key; });
-            }
-        };
-
-        constexpr ExpandedMapEntry *ExpandedMap::try_find_entry(IRIMapping const &key) {
-            auto i = std::ranges::find_if(entries, [&](auto const& t) { return t.key == key; });
-            if (i == entries.end()) {
-                return nullptr;
-            }
-            return &*i;
-        }
-        constexpr ExpandedMapEntry *ExpandedMap::try_find_keyword(std::string_view key) {
-            auto i = std::ranges::find_if(entries, [&](auto const& t)
-                { return t.key.type == IRIMappingType::Keyword && t.key.data == key; });
-            if (i == entries.end()) {
-                return nullptr;
-            }
-            return &*i;
-        }
-
-        struct StringifyResult {
-            std::string value;
-            std::string_view datatype;
-        };
-    }  // namespace json_ld
-
     namespace params {
+        struct ParseContextTermParams {
+            simdjson::ondemand::object local_context;
+            json_ld::Context &active_context; // NOLINT(*-avoid-const-or-ref-data-members)
+            json_ld::TermDefinition &term; // NOLINT(*-avoid-const-or-ref-data-members)
+            std::vector<json_ld::TermDefinition> const &previous_terms; // NOLINT(*-avoid-const-or-ref-data-members)
+            std::string_view base_iri;
+            bool is_protected = false;
+            bool override_protected = false;
+        };
+        struct ParseContextIRIExpansionParams {
+            json_ld::Context &active_context; // NOLINT(*-avoid-const-or-ref-data-members)
+            simdjson::ondemand::object local_context;
+            std::vector<json_ld::TermDefinition> const &previous_terms; // NOLINT(*-avoid-const-or-ref-data-members)
+        };
         struct ParseParams {
             simdjson::ondemand::value element;
-            json_ld::Context const &active_ctx;
+            json_ld::Context const &active_ctx; // NOLINT(*-avoid-const-or-ref-data-members)
             std::string_view base_iri;
-            json_ld::IRIMapping const &active_graph;
-            json_ld::IRIMapping const &active_subject;
-            json_ld::IRIMapping const &active_property;
+            json_ld::IRIMapping const &active_graph; // NOLINT(*-avoid-const-or-ref-data-members)
+            json_ld::IRIMapping const &active_subject; // NOLINT(*-avoid-const-or-ref-data-members)
+            json_ld::IRIMapping const &active_property; // NOLINT(*-avoid-const-or-ref-data-members)
             std::variant<std::monostate, json_ld::IRIMapping, Literal> *obj_out = nullptr;
             bool is_top_level = false;
             bool is_reverse = false;
@@ -260,7 +50,7 @@ namespace rdf4cpp::parser {
         state_type *state_;
         bool state_is_owned_;
         std::string json_data_;
-        uint64_t blank_node_index = 0;
+        uint64_t blank_node_index_ = 0;
 
         static constexpr bool any_of(std::string_view v, std::initializer_list<std::string_view> l) {
             for (auto const x : l) {
@@ -270,107 +60,8 @@ namespace rdf4cpp::parser {
             }
             return false;
         }
-        static constexpr bool is_keyword(std::string_view v) {
-            return any_of(v,
-                          {keyword_base,
-                           keyword_container,
-                           keyword_context,
-                           keyword_direction,
-                           keyword_graph,
-                           keyword_id,
-                           keyword_import,
-                           keyword_included,
-                           keyword_index,
-                           keyword_json,
-                           keyword_language,
-                           keyword_list,
-                           keyword_nest,
-                           keyword_none,
-                           keyword_prefix,
-                           keyword_propagate,
-                           keyword_protected,
-                           keyword_reverse,
-                           keyword_set,
-                           keyword_type,
-                           keyword_value,
-                           keyword_version,
-                           keyword_vocab});
-        }
-        static bool looks_like_keyword(std::string_view v);
 
-        template<typename T>
-        static auto try_get_field(simdjson::ondemand::object o, std::string_view key) {
-            T r{};
-            auto c = o[key].get(r);  // NOLINT(*-pro-bounds-avoid-unchecked-container-access)
-            return std::tuple(c, r);
-        }
-        template<typename T>
-        static auto try_get_field(simdjson::ondemand::array a, size_t key) {
-            T r{};
-            a.reset();
-            auto c = a.at(key).get(r);
-            return std::tuple(c, r);
-        }
-        template<typename T>
-        static auto try_get_field(simdjson::ondemand::object o, json_ld::KeyPath const &key) {
-            if (key.keys.empty()) {
-                if constexpr (std::same_as<T, simdjson::ondemand::object>) {
-                    return std::tuple(simdjson::SUCCESS, o);
-                }
-                else {
-                    return std::tuple(simdjson::INCORRECT_TYPE, T{});
-                }
-            }
-            simdjson::error_code c;
-            std::variant<simdjson::ondemand::array, simdjson::ondemand::object> e = o;
-            for (size_t i = 0; i < key.keys.size() - 1; ++i) {
-                const auto& next_key = key.keys[i + 1];  // NOLINT(*-pro-bounds-avoid-unchecked-container-access)
-                const auto& current_key = key.keys[i];  // NOLINT(*-pro-bounds-avoid-unchecked-container-access)
-                if (std::holds_alternative<size_t>(next_key)) {
-                    if (std::holds_alternative<size_t>(current_key)) {
-                        std::tie(c, e) = try_get_field<simdjson::ondemand::array>(std::get<simdjson::ondemand::array>(e), std::get<size_t>(current_key));
-                    }
-                    else {
-                        std::tie(c, e) = try_get_field<simdjson::ondemand::array>(std::get<simdjson::ondemand::object>(e), std::get<std::string>(current_key));
-                    }
-                }
-                else {
-                    if (std::holds_alternative<size_t>(current_key)) {
-                        std::tie(c, e) = try_get_field<simdjson::ondemand::object>(std::get<simdjson::ondemand::array>(e), std::get<size_t>(current_key));
-                    }
-                    else {
-                        std::tie(c, e) = try_get_field<simdjson::ondemand::object>(std::get<simdjson::ondemand::object>(e), std::get<std::string>(current_key));
-                    }
-                }
-                if (c != simdjson::SUCCESS) {
-                    return std::tuple(c, T{});
-                }
-            }
-            const auto& current_key = key.keys[key.keys.size() - 1];  // NOLINT(*-pro-bounds-avoid-unchecked-container-access)
-            if (std::holds_alternative<size_t>(current_key)) {
-                return try_get_field<T>(std::get<simdjson::ondemand::array>(e), std::get<size_t>(current_key));
-            }
-            else {
-                return try_get_field<T>(std::get<simdjson::ondemand::object>(e), std::get<std::string>(current_key));
-            }
-        }
-        template<typename T>
-        static std::tuple<simdjson::error_code, std::optional<T>> try_get_optional_field(simdjson::ondemand::object o,
-                                                                                         std::string_view key) {
-            simdjson::ondemand::value v;
-            auto c = o[key].get(v);  // NOLINT(*-pro-bounds-avoid-unchecked-container-access)
-            if (c != simdjson::SUCCESS) {
-                return std::tuple(c, std::nullopt);
-            }
-            if (v.is_null()) {
-                return std::tuple(c, std::nullopt);
-            }
-            T r{};
-            c = v.get(r);
-            return std::tuple(c, r);
-        }
-
-        error_type make_error(ParsingError::Type t, std::string msg);
+        [[nodiscard]] error_type make_error(ParsingError::Type t, std::string msg) const;
         json_ld::IRIMapping make_new_bn();
         nonstd::expected<IRI, error_type> make_iri(std::string_view i);
         nonstd::expected<IRI, error_type> make_iri(json_ld::IRIMapping const & m);
@@ -385,14 +76,7 @@ namespace rdf4cpp::parser {
                                                                      std::string_view base_iri,
                                                                      bool override_protected = false,
                                                                      bool propagate = true);
-        std::optional<error_type> parse_context_term(simdjson::ondemand::object local_context,
-                                                     json_ld::Context &active_context,
-                                                     json_ld::TermDefinition &term,
-                                                     json_ld::Context const &parent_context,
-                                                     std::vector<json_ld::TermDefinition> const & previous_terms,
-                                                     std::string_view base_iri,
-                                                     bool is_protected = false,
-                                                     bool override_protected = false);
+        std::optional<error_type> parse_context_term(params::ParseContextTermParams p);
 
         nonstd::expected<json_ld::Context, error_type> parse_local_context(simdjson::padded_string_view json,
                                                                      json_ld::Context const &active_context,
@@ -405,11 +89,8 @@ namespace rdf4cpp::parser {
                                                                         std::optional<std::string_view> value,
                                                                         bool document_relative,
                                                                         bool vocab,
-                                                                        json_ld::Context *local_context = nullptr,
-                                                                        std::optional<simdjson::ondemand::object> local_context_json
-                                                                        = std::nullopt,
                                                                         json_ld::TermDefinition const *ignore_local = nullptr,
-                                                                        std::vector<json_ld::TermDefinition> const * previous_terms = nullptr);
+                                                                        params::ParseContextIRIExpansionParams* parse_ctx = nullptr);
 
         nonstd::expected<json_ld::ExpandedValue, error_type> value_expansion(json_ld::Context const &active_conext,
                                                                              json_ld::IRIMapping const &active_property,
